# public/swagger.yml
openapi: "3.0.3"
info:
  title: "System Zarządzania Dostępem do Aparatury Naukowej"
  description: |
    System zarządzania czasem dostępu do aparatury naukowej pozwala na efektywne rezerwowanie urządzeń, monitorowanie dostępności oraz zarządzanie użytkownikami.
    Użytkownicy mogą rezerwować czas na urządzenia, sprawdzać dostępność w określonych terminach, a administratorzy mają możliwość zarządzania użytkownikami i kontrolowania zasobów.
    System wspiera przydzielanie czasu dostępu do aparatury, zarządzanie limitami użytkowników oraz generowanie raportów z wykorzystania urządzeń.
  version: "1.0.0"
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /bookings:
    get:
      summary: "Retrieve Bookings"
      description: "Fetches a list of bookings. Supports filtering by various parameters such as device ID, user ID, and date range."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/deviceId"
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
      tags:
        - Bookings
      responses:
        200:
          $ref: "#/components/responses/BOOKINGS_RETRIEVED"

        401:
          description: "Unauthorized - User not authenticated."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_AUTHENTICATED"

        500:
          description: "Internal Server Error - Unexpected error occurred."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UNEXPECTED_ERROR"
    post:
      summary: "Create Booking"
      description: "Creates a new booking. Requires authentication and validates input."
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        200:
          $ref: "#/components/responses/BOOKING_CREATED"

        400:
          description: "Bad Request - Invalid input or JSON"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/INVALID_DATA"
                  - $ref: "#/components/schemas/INVALID_JSON_BODY"
                  - $ref: "#/components/schemas/NO_DATA_PROVIDED"
        401:
          description: "Unauthorized - User not authenticated."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_AUTHENTICATED"

        403:
          $ref: "#/components/responses/MONTHLY_LIMIT_EXCEEDED"

        404:
          description: "Not Found - User or Device not found."
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/USER_NOT_FOUND"
                  - $ref: "#/components/schemas/DEVICE_NOT_FOUND"
        409:
          $ref: "#/components/responses/DEVICE_UNAVAILABLE"

        500:
          description: "Internal Server Error - Unexpected error occurred."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UNEXPECTED_ERROR"
...
  /devices/{deviceId}/bookings:
    get:
      summary: "Retrieve bookings for a specific device"
      description: "Fetches a paginated list of bookings for a specific device within an optional date range and with optional filters for the user ID. Requires user authentication."
      parameters:
        - $ref: "#/components/parameters/deviceIdRoute"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
      tags:
        - Devices
        - Bookings
      responses:
        200:
          $ref: "#/components/responses/BOOKINGS_RETRIEVED"

        401:
          description: "Unauthorized - The user must be authenticated to access this endpoint."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_AUTHENTICATED"

        404:
          description: "Device Not Found - No device exists with the provided ID."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DEVICE_NOT_FOUND"
        500:
          description: "Internal Server Error - An unexpected issue occurred while processing the request."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UNEXPECTED_ERROR"
...
  /me/bookings:
    get:
      summary: "Retrieve bookings for the authenticated user"
      description: "Fetches a paginated list of bookings made by the authenticated user. Supports filtering by various parameters, including device ID and date range."
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/deviceId"
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
      tags:
        - Self
        - Bookings
      responses:
        200:
          $ref: "#/components/responses/BOOKINGS_RETRIEVED"

        401:
          description: "Unauthorized - The user must be authenticated to access this endpoint."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_AUTHENTICATED"

        500:
          description: "Internal Server Error - An unexpected issue occurred while processing the request."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UNEXPECTED_ERROR"
...
  /users/{userId}/bookings:
    get:
      summary: "Retrieve Bookings for a Specific User"
      description: "Fetches a list of bookings for a specific user. The user must be authenticated and have sufficient permissions to view this data. Supports pagination and optional filters based on device and date range."
      parameters:
        - $ref: "#/components/parameters/userIdRoute"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/deviceId"
        - $ref: "#/components/parameters/startDate"
        - $ref: "#/components/parameters/endDate"
      tags:
        - Users
        - Bookings
      responses:
        200:
          $ref: "#/components/responses/BOOKINGS_RETRIEVED"

        401:
          description: "Unauthorized - The user must be authenticated to access this endpoint."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_AUTHENTICATED"

        403:
          $ref: "#/components/responses/INSUFFICIENT_PERMISSIONS"

        404:
          description: "Not Found - The user with the specified ID does not exist."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/USER_NOT_FOUND"

        500:
          description: "Internal Server Error - An unexpected error occurred while processing the request."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UNEXPECTED_ERROR"
...
components:
  parameters:
    page:
      name: page
      in: query
      description: "The page number to retrieve."
      required: false
      schema:
        type: integer
        example: 1
    limit:
      name: limit
      in: query
      description: "The number of results per page."
      required: false
      schema:
        type: integer
        example: 10
    deviceId:
      name: deviceId
      in: query
      description: "Filter for specific deviceId."
      required: false
      schema:
        type: string
        example: abc123
    deviceIdRoute:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        example: abc123
...
    userId:
      name: userId
      in: query
      description: "Filter for specific userId."
      required: false
      schema:
        type: string
        example: abc123
    userIdRoute:
      name: userId
      in: path
      required: true
      schema:
        type: string
        example: abc123
...
    startDate:
      name: startDate
      in: query
      description: "Filter for records after specified date"
      required: false
      schema:
        type: string
        example: 2024-11-25T17:00:00.180Z
    endDate:
      name: endDate
      in: query
      description: "Filter for records before specified date"
      required: false
      schema:
        type: string
        example: 2024-11-28T17:00:00.180Z
  schemas:
...
    CreateBookingRequest:
      type: object
      required:
        - userId
        - deviceId
        - startTime
        - endTime
      properties:
        userId:
          type: string
          description: "The ID of the user making the booking."
        deviceId:
          type: string
          description: "The ID of the device being booked."
        startTime:
          type: string
          format: date-time
          description: "The start time for the booking."
        endTime:
          type: string
          format: date-time
          description: "The end time for the booking."
...
    BasicUser:
      type: object
      properties:
        id:
          type: string
          example: "12345"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "johndoe@example.com"
...
    BasicDevice:
      type: object
      properties:
        id:
          type: string
          example: "dev123"
        name:
          type: string
          example: "Microscope 1"
        building:
          type: string
          example: "Building A"
        room:
          type: string
          example: "Room 101"
...
    DetailedBooking:
      type: object
      properties:
        id:
          type: string
          example: "bkng123"
        user:
          $ref: "#/components/schemas/BasicUser"
        device:
          $ref: "#/components/schemas/BasicDevice"
        startTime:
          type: string
          format: date-time
          example: "2024-11-25T14:30:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-11-25T16:30:00Z"
        isCanceled:
          type: boolean
          example: false
        duration:
          type: integer
          example: 2
        updatedAt:
          type: string
          format: date-time
          example: "2024-11-25T14:30:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-11-25T14:30:00Z"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 5
        totalCount:
          type: integer
          example: 50
    INVALID_JSON_BODY:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "The request body is not properly formatted as JSON. Please ensure the body is valid JSON."
        code:
          type: string
          example: "INVALID_JSON_BODY"
...
    NO_DATA_PROVIDED:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "No data was provided in the request body. Please provide the necessary information and try again."
        code:
          type: string
          example: "NO_DATA_PROVIDED"
    INVALID_DATA:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "The request data is invalid or missing required fields. Please check the input and try again."
        details:
          type: array
          description: "Detailed information about invalid fields."
          items:
            type: object
            properties:
              field:
                type: string
                description: "The name of the field that caused the validation error."
                example: "email"
              message:
                type: string
                description: "The error message for the invalid field."
                example: "'email' is required and must be a valid string."
        code:
          type: string
          example: "INVALID_DATA"
...
    USER_NOT_AUTHENTICATED:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "Authentication is required to access this resource. Please log in and try again."
        code:
          type: string
          example: "USER_NOT_AUTHENTICATED"
...
    UNEXPECTED_ERROR:
      type: object
      properties:
        error:
          type: string
          example: "Internal Server Error"
        message:
          type: string
          example: "An unexpected error occurred. Please try again later or contact admin if the issue persists."
        code:
          type: string
          example: "UNEXPECTED_ERROR"
    USER_NOT_FOUND:
      type: object
      properties:
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "No user found with the specified ID. Please verify the ID and try again."
        code:
          type: string
          example: "USER_NOT_FOUND"
    DEVICE_NOT_FOUND:
      type: object
      properties:
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "No device found with the specified ID. Please verify the ID and try again."
        code:
          type: string
          example: "DEVICE_NOT_FOUND"
...
  responses:
...
    BOOKING_CREATED:
      description: "Booking created successfully."
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Booking created successfully."
              data:
                $ref: "#/components/schemas/DetailedBooking"
              code:
                type: string
                example: "BOOKING_CREATED"
...
    BOOKINGS_RETRIEVED:
      description: "List of bookings retrieved successfully."
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Bookings retrieved successfully."
              data:
                type: array
                items:
                  $ref: "#/components/schemas/DetailedBooking"
              pagination:
                $ref: "#/components/schemas/Pagination"
              code:
                type: string
                example: "BOOKINGS_RETRIEVED"
...
    DEVICE_UNAVAILABLE:
      description: "The device is already reserved during the specified time."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Conflict"
              message:
                type: string
                example: "The selected device is already reserved during the specified time range. Please choose a different time or device."
              code:
                type: string
                example: "DEVICE_UNAVAILABLE"
    MONTHLY_LIMIT_EXCEEDED:
      description: "User's monthly limit has been exceeded."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Limit Exceeded"
              message:
                type: string
                example: "The user's monthly usage limit has been reached. Please try booking on a different day or adjust the booking duration."
              code:
                type: string
                example: "MONTHLY_LIMIT_EXCEEDED"
    INSUFFICIENT_PERMISSIONS:
      description: "User does not have sufficient permissions."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Forbidden"
              message:
                type: string
                example: "You do not have sufficient permissions to perform this action."
              code:
                type: string
                example: "INSUFFICIENT_PERMISSIONS"
...